Reorder files for various front-end components
The way we had our Angular components laid out before was incredibly messy. More info in the pull request on what the exact changes are that this commit makes.
